Refactor taxon search hook
